Address runtime review feedback
